refactor: configs as envsConfig
