+++ src/dev/mocks/mockedModules.ts
@@ -1,5 +1,5 @@
 import { v4 as uuidV4 } from 'uuid';
-import configs from 'src/modules/core/configs/configs.config';
+import envsConfig from 'src/modules/core/configs/envs.config';
 import { LoggerInterface } from 'src/modules/core/logging/logger';
 
 
@@ -7,7 +7,7 @@ export const configServiceMock = {
 	get: (propertyPath?: string): any => {
 		if (propertyPath) {
 			const splitedPaths = propertyPath.split('.');
-			let scopedProperty: any = configs();
+			let scopedProperty: any = envsConfig();
 
 			for (const scopedPath of splitedPaths) {
 				if (scopedPath.length)
@@ -16,7 +16,7 @@ export const configServiceMock = {
 
 			return scopedProperty;
 		} else
-			return configs();
+			return envsConfig();
 	},
 };
