+++ src/modules/core/infra/providers/RestMockedService.provider.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@nestjs/common';
 import { ConfigService } from '@nestjs/config';
-import { ConfigsInterface } from '@core/configs/configs.config';
+import { ConfigsInterface } from '@core/configs/envs.config';
 import AbstractRestClient from '@core/infra/integration/rest/AbstractRestClient.client';
 import LoggerService from '@core/logging/Logger.service';
 import catchError from '@common/utils/externalErrorParser.util';
