+++ src/modules/events/queue/producers/EventsQueue.producer.ts
@@ -5,7 +5,7 @@ import SqsClient from '@core/infra/integration/aws/Sqs.client';
 import { SINGLETON_LOGGER_PROVIDER, LoggerProviderInterface } from '@core/logging/Logger.service';
 import { LoggerInterface } from '@core/logging/logger';
 import CryptographyService from '@core/security/Cryptography.service';
-import { ConfigsInterface } from '@core/configs/configs.config';
+import { ConfigsInterface } from '@core/configs/envs.config';
 import { EventSchemaInterface } from '@events/queue/handlers/schemas/event.schema';
